+++ HISTORY.rst
@@ -12,6 +12,7 @@ History
     * Increase code coverage (ericof).
     * Fix Workflow imports (ericof).
     * New base resource to return a paginate collection from a custom plain sql query (rudaporto).
+    * Refactor all mocks used for resource tests (rudaporto).
 
 2.0.6 (2017-07-27)
 ------------------